Adding management samples
