+++ samples/Management/CreateTopicWithSubscriptionsAndRules/SubscriptionFilterCli.py
