+++ samples/Management/CreateTopicWithSubscriptionsAndRules/SubscriptionFilterPS.ps1
