+++ samples/Management/GeoDR/SBGeoDR2/SBGeoDR2/App.config
