+++ samples/Management/GeoDR/SBGeoDR2/SBGeoDR2/AssemblyInfo.cs
