+++ samples/Management/GeoDR/SBGeoDR2/SBGeoDR2/Properties/AssemblyInfo.cs
