+++ samples/Management/GeoDR/SBGeoDR2/SBGeoDR2/SBGeoDR2.csproj
