+++ samples/Management/GeoDR/SBGeoDR2/SBGeoDR2/packages.config
