+++ samples/Management/GeoDR/SBGeoDR2/SBGeoDR_existing_namespace_name/Properties/AssemblyInfo.cs
