+++ samples/Management/GeoDR/SBGeoDR2/SBGeoDR_existing_namespace_name/SBGeoDR2_existing_namespace_name.sln
