+++ samples/Management/GeoDR/SBGeoDR2/SBGeoDR_existing_namespace_name/packages.config
