+++ samples/Management/GeoDR/TestGeoDR/ConsoleApp1/App.config
