+++ samples/Management/GeoDR/TestGeoDR/ConsoleApp1/AssemblyInfo.cs
