+++ samples/Management/GeoDR/TestGeoDR/ConsoleApp1/Program.cs
