+++ samples/Management/GeoDR/TestGeoDR/ConsoleApp1/Properties/AssemblyInfo.cs
