+++ samples/Management/GeoDR/TestGeoDR/ConsoleApp1/TestGeoDR.csproj
