+++ samples/Management/GeoDR/TestGeoDR/ConsoleApp1/TestGeoDR.sln
