+++ samples/Management/GeoDR/TestGeoDR/ConsoleApp1/packages.config
